Move Context.h up into rpc/.
