+++ Builds/VisualStudio2013/RippleD.vcxproj.filters
@@ -4119,6 +4119,9 @@
     <ClCompile Include="..\..\src\ripple\resource\tests\Logic.test.cpp">
       <Filter>ripple\resource\tests</Filter>
     </ClCompile>
+    <ClInclude Include="..\..\src\ripple\rpc\Context.h">
+      <Filter>ripple\rpc</Filter>
+    </ClInclude>
     <ClInclude Include="..\..\src\ripple\rpc\Coroutine.h">
       <Filter>ripple\rpc</Filter>
     </ClInclude>
@@ -4309,9 +4312,6 @@
       <Filter>ripple\rpc\impl</Filter>
     </ClCompile>
     <ClInclude Include="..\..\src\ripple\rpc\impl\Accounts.h">
-      <Filter>ripple\rpc\impl</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\ripple\rpc\impl\Context.h">
       <Filter>ripple\rpc\impl</Filter>
     </ClInclude>
     <ClCompile Include="..\..\src\ripple\rpc\impl\Coroutine.cpp">
